# docs/openapi.json
{
  "components": {
    "responses": {
      "Error": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "An error."
      }
    },
    "schemas": {
      "Admin": {
        "properties": {
          "auth_source": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "disabled": {
            "type": "boolean"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "last_login_at": {
            "format": "date-time",
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "scope": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "totp_enabled": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "CreateAdminRequest": {
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ListOfAdmin": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Admin"
            },
            "type": "array"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "totp_code": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "SessionResponse": {
        "properties": {
          "admin": {
            "$ref": "#/components/schemas/Admin"
          },
          "agent_download_url": {
            "type": "string"
          },
          "csrf_token": {
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "type": "string"
          },
          "signing_required": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "SetupResponse": {
        "properties": {
          "local_login": {
            "type": "boolean"
          },
          "needs_setup": {
            "type": "boolean"
          },
          "sso": {
            "$ref": "#/components/schemas/SsoSetup"
          }
        },
        "type": "object"
      },
...
      "SsoSetup": {
        "properties": {
          "display_name": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "session": {
        "description": "A console session. Changes also need the X-CSRF-Token header.",
        "in": "cookie",
        "name": "retune_session",
        "type": "apiKey"
      },
      "token": {
        "description": "An API token, rtk_…",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "The API the Retune console uses. Sign in with a session (the retune_session cookie, with the X-CSRF-Token header on every change) or send an API token as `Authorization: Bearer rtk_…`. Each operation's `x-retune-access` names who may call it. Errors are `{\"code\": \"…\", \"message\": \"…\"}`.",
    "title": "Retune admin API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admins": {
      "get": {
        "description": "Signed in to the console, not limited to some devices.",
        "operationId": "getAdmins",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListOfAdmin"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "List admins",
        "tags": [
          "Administration"
        ],
        "x-retune-access": "admin"
      },
      "post": {
        "description": "The admin role, signed in to the console, not limited to some devices.",
        "operationId": "postAdmins",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Admin"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Create an admin",
        "tags": [
          "Administration"
        ],
        "x-retune-access": "admin+write"
      }
    },
...
    "/oidc/callback": {
      "get": {
        "description": "The identity provider redirects here. Sets the session cookie and redirects to the console.\n\nNo sign-in.",
        "operationId": "getOidcCallback",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "error",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "error_description",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Found"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [],
        "summary": "Finish single sign-on",
        "tags": [
          "Session"
        ],
        "x-retune-access": "public"
      }
    },
    "/oidc/start": {
      "get": {
        "description": "Redirects to the identity provider. 404 if SSO is not configured.\n\nNo sign-in.",
        "operationId": "getOidcStart",
        "responses": {
          "302": {
            "description": "Found"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [],
        "summary": "Start single sign-on",
        "tags": [
          "Session"
        ],
        "x-retune-access": "public"
      }
    },
...
    "/session": {
      "delete": {
        "description": "A person signed in to the console; not an API token.",
        "operationId": "deleteSession",
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Sign out",
        "tags": [
          "Session"
        ],
        "x-retune-access": "session"
      },
      "get": {
        "description": "A person signed in to the console; not an API token.",
        "operationId": "getSession",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "The signed-in admin",
        "tags": [
          "Session"
        ],
        "x-retune-access": "session"
      },
      "post": {
        "description": "Sets the retune_session cookie. The response carries the CSRF token to send as X-CSRF-Token.\n\nNo sign-in.",
        "operationId": "postSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [],
        "summary": "Sign in with a password (and authenticator code)",
        "tags": [
          "Session"
        ],
        "x-retune-access": "public"
      }
    },
    "/setup": {
      "get": {
        "description": "No sign-in.",
        "operationId": "getSetup",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SetupResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [],
        "summary": "Whether first-run setup is needed, and how people sign in",
        "tags": [
          "Session"
        ],
        "x-retune-access": "public"
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/admin/v1"
    }
  ]
}